requirements.txt
python-telegram-bot==20.3
emoji
requests
python-dotenv